% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_rubins.R
\name{combine_rubins}
\alias{combine_rubins}
\title{combine_rubins}
\usage{
combine_rubins(
  DGE,
  model_results,
  voom_formula,
  covariate = NULL,
  robust = FALSE,
  winsor.tail.p = c(0.05, 0.1)
)
}
\arguments{
\item{DGE}{A DGEList object.}

\item{model_results}{Output from limmavoom_imputed_datalist or limmavoom_imputed_data_list_parallel.}

\item{voom_formula}{Formula for design matrix. Should be same formula used to create model_results.}

\item{covariate}{See limma squeezeVar function.}

\item{robust}{See limma squeezeVar function.}

\item{winsor.tail.p}{See limma squeezeVar function.}
}
\value{
Dataframe with one row per gene containing coefficients standard errors, degrees of freedom, t-statistics, P-Values, and adjusted P-values from the limma-voom pipeline.
}
\description{
Combines results from each imputed dataset using Rubin's rules.
}
\examples{
intervals <- get_gene_bin_intervals(DGE, data, n = 10)
gene_bin_impute <- impute_by_gene_bin(data,
    intervals,
    DGE,
    n = 3
)
coef_se <- limmavoom_imputed_data_list(
    gene_intervals = intervals,
    DGE = DGE,
    imputed_data_list = gene_bin_impute,
    n = 3,
    voom_formula = "~x + y + z + a + b",
    predictor = "x"
)

final_res <- combine_rubins(
    DGE = DGE,
    model_results = coef_se,
    voom_formula = "~x + y + z + a + b"
)
}
