% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_rubins_pca.R
\name{combine_rubins_pca}
\alias{combine_rubins_pca}
\title{combine_rubins_pca}
\usage{
combine_rubins_pca(gene, results_final, avg_df, m)
}
\arguments{
\item{gene}{The number gene in the DGE list.}

\item{results_final}{List with one element per gene storing coefficients and standard errors from limma-voom analysis.}

\item{avg_df}{The average df.total from limma-voom analysis across the m imputed datasets.}

\item{m}{Number of imputed data sets.}
}
\value{
Coefficient, standard error, and p-value combined across m imputed datasets.
}
\description{
Combines results from each imputed dataset using Rubin's rules.
}
\keyword{internal}
