% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Simulated dataset}
\format{
\subsection{\code{example_data}}{

data frame with 500 rows and 5 variables
\describe{
\item{x}{continuous normally distributed}
\item{y}{continuous normally distributed}
\item{z}{continuous normally distributed}
\item{a}{binary}
\item{b}{binary}
...
}
}
}
\usage{
data(example_data)
}
\value{
Tibble with 500 rows of data for variables x, y, and z
}
\description{
The exact code used to generate these data are found in the Example_Data_for_RNAseqCovarImpute
vignette. In short, \code{example_data} contains 500 rows with data for variables x, y, and z, which
are continuous normally distributed, and a and b, which are binary variables. Missigness
was simulated for all variables other than x such that a complete case analysis would drop
24.2\% of participants. \code{example_DGE} contains random count data from the Poisson distribution
for 500 made up genes, ENS1-ENS500
}
\examples{
data(example_data)
}
\keyword{datasets}
