% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_by_gene_bin_parallel.R
\name{impute_by_gene_bin_parallel}
\alias{impute_by_gene_bin_parallel}
\title{impute_by_gene_bin_parallel}
\usage{
impute_by_gene_bin_parallel(data, intervals, DGE, n, cores)
}
\arguments{
\item{data}{Sample data with one row per sample. Sample row order should match the col order in the DGEList.}

\item{intervals}{Output from get_gene_bin_intervals function. A dataframe where each row contains the start (first col) and end (second col) values for each gene bin interval.}

\item{DGE}{A DGEList object.}

\item{n}{Number of imputed data sets.}

\item{cores}{Number of cores to run in parallel using 'PSOCK' back-end implemented with doParallel.}
}
\value{
A list of sets of n imputed datasets, one per gene bin.
}
\description{
Loops through DGE list using the gene bin intervals from the
"get_gene_bin_intervals" function and makes imputed datasets. For instance,
if n = 100 and intervals contains 200 gene bin intervals, output will be a list of
200 sets of 100 imputed datasets. Each of the 200 sets are imputed
using only the genes in one gene bin.
}
\examples{
intervals <- get_gene_bin_intervals(DGE, data, n = 10)
gene_bin_impute <- impute_by_gene_bin_parallel(data,
    intervals,
    DGE,
    n = 3,
    cores = 2
)
coef_se <- limmavoom_imputed_data_list_parallel(
    gene_intervals = intervals,
    DGE = DGE,
    imputed_data_list = gene_bin_impute,
    n = 3,
    voom_formula = "~x + y + z + a + b",
    predictor = "x",
    cores = 2
)

final_res <- combine_rubins(
    DGE = DGE,
    model_results = coef_se,
    voom_formula = "~x + y + z + a + b"
)
}
