% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_by_gene_bin_helper.R
\name{impute_gene_bin_helper}
\alias{impute_gene_bin_helper}
\title{impute_by_gene_bin_helper}
\usage{
impute_gene_bin_helper(i, intervals, cpm_all, data, m, maxit)
}
\arguments{
\item{intervals}{Output from get_gene_bin_intervals function. A dataframe where each row contains the start (first col) and end (second col) values for each gene bin interval.}

\item{data}{Sample data with one row per sample. Sample row order should match the col order in the DGEList.}

\item{m}{Number of imputed data sets.}

\item{maxit}{Used by mice function.}

\item{DGE}{A DGEList object.}

\item{param}{Arguments passed to BiocParallel::bpparam()}
}
\value{
A list of sets of n imputed datasets, one per gene bin.
}
\description{
Loops through DGE list using the gene bin intervals from the
"get_gene_bin_intervals" function and makes imputed datasets. For instance,
if n = 100 and intervals contains 200 gene bin intervals, output will be a list of
200 sets of 100 imputed datasets. Each of the 200 sets are imputed
using only the genes in one gene bin.
}
\keyword{internal}
