% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmavoom_imputed_data_list_helper.R
\name{limmavoom_imputed_data_list_helper}
\alias{limmavoom_imputed_data_list_helper}
\title{limmavoom_imputed_data_list_helper}
\usage{
limmavoom_imputed_data_list_helper(
  gene_bin,
  gene_intervals,
  DGE,
  imputed_data_list,
  m,
  voom_formula,
  predictor,
  sx_sy
)
}
\arguments{
\item{gene_intervals}{Output from get_gene_bin_intervals function. A dataframe where each row contains the start (first col) and end (second col) values for each gene bin interval.}

\item{DGE}{A DGEList object.}

\item{imputed_data_list}{Output from impute_by_gene_bin.}

\item{m}{Number of imputed data sets.}

\item{voom_formula}{Formula for design matrix.}

\item{predictor}{Independent variable of interest. Must be a variable in voom_formula.}
}
\value{
A dataframe with coefficient, standard error, sigma, and residual degrees of freedom values from limma-voom gene expression analysis. One row per gene and one set of values per imputed dataset.
}
\description{
Loops through the imputed data list (output from "impute_by_gene_bin" function)
and runs limma-voom RNA seq analysis.
}
\keyword{internal}
