% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmavoom_imputed_data_pca_helper.R
\name{limmavoom_imputed_data_pca_helper}
\alias{limmavoom_imputed_data_pca_helper}
\title{limmavoom_imputed_data_pca_helper}
\usage{
limmavoom_imputed_data_pca_helper(i, imp, DGE, voom_formula)
}
\arguments{
\item{i}{Number from 1-m}

\item{imp}{Imputed data from mice (mids object)}

\item{DGE}{A DGEList object.}

\item{voom_formula}{Formula for design matrix.}
}
\value{
A list with coefficients, standard errors, and total degrees of freedom values from limma-voom gene expression analysis.
}
\description{
Runs limma-voom pipeline on one of the m imputed datasets and returns the coefficients,
Bayesian moderated standard errors and degrees of freedom.
This function used while looping through the m imputed datasets.
}
\keyword{internal}
