% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowess_all_gene_bins.R
\name{lowess_all_gene_bins}
\alias{lowess_all_gene_bins}
\title{lowess_all_gene_bins}
\usage{
lowess_all_gene_bins(
  gene_intervals,
  DGE,
  imputed_data_list,
  m,
  voom_formula,
  predictor
)
}
\description{
Loops through all bins and all M imputations, prepares DGE and design to run
voom_sx_sy, which fits gene-wise linear models and extracts log count size (sx)
and sqrt resudual standard deviations (sy) to make the lowess curve
}
