% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voom_sx_sy.R
\name{voom_sx_sy}
\alias{voom_sx_sy}
\title{voom_sx_sy}
\usage{
voom_sx_sy(
  counts,
  design = NULL,
  lib.size = NULL,
  normalize.method = "none",
  block = NULL,
  correlation = NULL,
  weights = NULL,
  span = 0.5,
  plot = FALSE,
  save.plot = FALSE,
  lib.size.all
)
}
\value{
Tibble with one col for sx and one for sy for lowess function.
}
\description{
Modified voom function used by limma_voom-imputed_data_list function.
Allows input of bins of outcome genes while still accounting for the total library size
of all outcome genes, as the total library size is needed to calculate log-cpm values.
Returns just the sx and sy values needed for lowess curve.
}
\examples{
data(example_data)
data(example_DGE)
intervals <- get_gene_bin_intervals(example_DGE, example_data, n = 10)
gene_bin_impute <- impute_by_gene_bin(example_data,
    intervals,
    example_DGE,
    m = 2,
    param = SerialParam()
)
coef_se <- limmavoom_imputed_data_list(
    gene_intervals = intervals,
    DGE = example_DGE,
    imputed_data_list = gene_bin_impute,
    m = 2,
    voom_formula = "~x + y + z + a + b",
    predictor = "x",
    param = SerialParam()
)

final_res <- combine_rubins(
    DGE = example_DGE,
    model_results = coef_se,
    voom_formula = "~x + y + z + a + b"
)
}
\keyword{internal}
